LayoutTests/fast/forms/date-multiple-fields/date-multiple-fields-readonly-subfield-expected.txt

Sub-fields in input[type=date] should be read-only in some cases. This requires window.internals.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


createDateInput argument order: min, max, value

Year field:
PASS isReadOnlyField(createDateInput("", "", ""), pseudoYear) is false
PASS isReadOnlyField(createDateInput("2012-12-17", "", ""), pseudoYear) is false
PASS isReadOnlyField(createDateInput("", "2012-12-17", ""), pseudoYear) is false
PASS isReadOnlyField(createDateInput("", "", "2012-12-17"), pseudoYear) is false
PASS isReadOnlyField(createDateInput("2012-01-01", "2012-12-31", ""), pseudoYear) is true
PASS isReadOnlyField(createDateInput("2012-01-01", "2013-12-31", ""), pseudoYear) is false
PASS isReadOnlyField(createDateInput("2012-01-01", "2012-12-31", "2012-05-30"), pseudoYear) is true
PASS isReadOnlyField(createDateInput("2012-01-01", "2012-12-31", "2013-01-01"), pseudoYear) is false
Month field:
PASS isReadOnlyField(createDateInput("", "", ""), pseudoMonth) is false
PASS isReadOnlyField(createDateInput("2012-12-17", "", ""), pseudoMonth) is false
PASS isReadOnlyField(createDateInput("", "2012-12-17", ""), pseudoMonth) is false
PASS isReadOnlyField(createDateInput("", "", "2012-12-17"), pseudoMonth) is false
PASS isReadOnlyField(createDateInput("2012-12-01", "2012-12-31", ""), pseudoMonth) is true
PASS isReadOnlyField(createDateInput("2012-11-01", "2013-12-31", ""), pseudoMonth) is false
PASS isReadOnlyField(createDateInput("2012-12-01", "2013-12-31", ""), pseudoMonth) is false
PASS isReadOnlyField(createDateInput("2012-12-01", "2012-12-31", "2012-12-15"), pseudoMonth) is true
PASS isReadOnlyField(createDateInput("2012-12-01", "2012-12-31", "2012-11-30"), pseudoMonth) is false
Day field:
PASS isReadOnlyField(createDateInput("", "", ""), pseudoDay) is false
PASS isReadOnlyField(createDateInput("2012-12-17", "", ""), pseudoDay) is false
PASS isReadOnlyField(createDateInput("", "2012-12-17", ""), pseudoDay) is false
PASS isReadOnlyField(createDateInput("", "", "2012-12-17"), pseudoDay) is false
PASS isReadOnlyField(createDateInput("2012-12-17", "2012-12-17", ""), pseudoDay) is true
PASS isReadOnlyField(createDateInput("2012-12-17", "2013-12-18", ""), pseudoDay) is false
PASS isReadOnlyField(createDateInput("2012-11-17", "2012-12-17", ""), pseudoDay) is false
PASS isReadOnlyField(createDateInput("2012-12-17", "2013-12-17", ""), pseudoDay) is false
PASS isReadOnlyField(createDateInput("2012-12-17", "2012-12-17", "2012-12-17"), pseudoDay) is true
PASS isReadOnlyField(createDateInput("2012-12-17", "2012-12-17", "2012-12-18"), pseudoDay) is false
PASS successfullyParsed is true

TEST COMPLETE
